Show progress bar with parallel sampling
This commit enables a progress bar when using `psample`. This done by
using a progress bar for the first job. Although the information is for
only a single job, the progress of other jobs should usually be very
similar (and it is better than having no feedback on the state of the
sampler).

This closes #374.