Depend on Theano's master branch (not v0.6)
`as_op` was introduced after the 0.6 release of Theano (commit
c5b999ed).

See faf560d and #389 for an explanation of why '<=' is used instead of
'=='. As discussed in #389, this won't trigger a Theano update if the
user already has it installed, so previous Theano versions will have to
be removed first. Of course, this will only matter if the `as_op`
functionality is being used.

Fixes #613.