Metaclass-based solution.
Alternative approach to solving the context stack issues.
The previous version was a dead end, because there was no effective way to check the appropriateness of the classes on a single context stack.
So in this version, I split the context stacks into two, one for the pm.Models, and one for pm.distributions.distributions._DrawValuesContext.
This works better, but involved replacing the Context *class* with the ContextMeta parameterized metaclass.