Add .random method to MvGaussianRandomWalk (#4388)
* Implement random for MvGaussianRandomWalk

Implements the random method for MvGaussianRandomWalk,
partially fixing #4337.

* Update docstring for GaussianRandomWalk._random()

* Add example to random docstring

* Modify MvGaussianRandomWalk.random

Improves the implementation by using MvNormal.random as suggested by
@Sayam753. Also updated its docstring to add more examples.

* Add TestMvGaussianRandomWalk

* Pass entire shape to MvNormal

This commit rewrites the random method to pass the entire shape
into MvNormal. MvGaussianRandomWalk also now supports an integer shape
that must match the dimensionality of mu. In this case it is assumed
that there are no time steps, and a random sample from the multivariate
normal distribution will be returned.

* Fix rebase issue

* Update pymc3/distributions/timeseries.py

Remove equality since the shape parameter cannot have more than 2 dimensions.

Co-authored-by: Sayam Kumar <sayamkumar753@yahoo.in>

* Update comment pymc3/distributions/timeseries.py

Co-authored-by: Sayam Kumar <sayamkumar753@yahoo.in>

* Update pymc3/distributions/timeseries.py

Adds a more explicit conditional on the time_axis statement.

Co-authored-by: Sayam Kumar <sayamkumar753@yahoo.in>

* Fix syntax error

* Add tests for MvGaussianRandomWalk with RV params

Added tests for both chol and cov to be random variables.

* Improve comment clarity in MvNormal._random

Moves the brittle comment in MvRandom._random's docstring to the actual location in the code for clarity.

* Update docstring formatting

* Update RELEASE-NOTES.md

Co-authored-by: Sayam Kumar <sayamkumar753@yahoo.in>