Simplify control flow of when initial points are determined
The initial point is now determined exactly once in the control flow:
+ By `init_nuts` (initvals replace init results).
+ In `sample`, if the above does not apply or fails.

Lower-level sampling functions now require the `start` kwarg to be a
complete dictionary of numeric initial values for all free variables.

The initial points for _each_ chain is checked for shape and logp inf/nan
once in `sample`, even if they may be identical for all chains.

Co-authored-by: Osvaldo Martin <aloctavodia@gmail.com>