Tweak prior in `test_deterministic_of_observed`
This test was failing due to the logp assertions introduced by aeppl (see https://github.com/aesara-devs/aeppl/issues/84), which combined with the strongly informative prior away from the likelihood led to underflow of sigma to 0, when using float32, and a subsequent AssertionError. Since this test is only concerned with the accuracy of the deterministic I have simply made the prior less strict.