Drop Python 3.7 and start testing on 3.10
Tests were redistributed between Python versions such that all of
3.8, 3.9, 3.10 are covered.

The exact `mkl-service` version was unpinned and `mkl` was removed
because the combination of pinned versions didn't support Python 3.10.
The `mkl` package is already a dependency of `mkl-service` and
because `mkl-service` specifies tight version limits for `mkl`,
we shouldn't pin it ourselves.

Closes #5209